requirements.txt
# Core Dependencies
pyspark==3.5.0
kafka-python==2.0.2
findspark==2.0.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
structlog==23.2.0

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0 